+++ hbase-endpoint/src/test/java/org/apache/hadoop/hbase/coprocessor/TestCoprocessorTableEndpoint.java
@@ -159,12 +159,12 @@ public class TestCoprocessorTableEndpoint {
     try {
       Map<byte[], Long> results =
         sum(table, TEST_FAMILY, TEST_QUALIFIER, ROWS[0], ROWS[ROWS.length - 1]);
-      int sumResult = 0;
-      int expectedResult = 0;
+      long sumResult = 0;
+      long expectedResult = 0;
       for (Map.Entry<byte[], Long> e : results.entrySet()) {
         sumResult += e.getValue();
       }
-      for (int i = 0; i < ROWSIZE; i++) {
+      for (long i = 0; i < ROWSIZE; i++) {
         expectedResult += i;
       }
       assertEquals("Invalid result", expectedResult, sumResult);
